+++ DependencyInjection/NelmioSecurityExtension.php
@@ -134,6 +134,7 @@ class NelmioSecurityExtension extends Extension
             $container->setParameter('nelmio_security.forced_ssl.hsts_preload', $config['forced_ssl']['hsts_preload']);
             $container->setParameter('nelmio_security.forced_ssl.whitelist', $config['forced_ssl']['whitelist']);
             $container->setParameter('nelmio_security.forced_ssl.hosts', $config['forced_ssl']['hosts']);
+            $container->setParameter('nelmio_security.forced_ssl.x_forwarded_proto', $config['forced_ssl']['x_forwarded_proto']);
         }
     }
 }